requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
python-dotenv==1.0.0
wtforms
flask_wtf
neo4j
py2neo
requests
psutil